requirements.txt
amqp==2.1.3
billiard==3.5.0.2
celery==4.0.1
Django==1.9
django-braces==1.9.0
django-cors-headers==1.1.0
django-filter==0.12.0
django-js-reverse==0.7.2
django-model-utils==2.4
django-mptt==0.8.6
django-multiuploader==0.2.9
django-oauth-toolkit==0.10.0
djangorestframework==3.4.6
fake-factory==0.7.2
ipaddress==1.0.17
ipdb==0.8
ipython==3.1.0
isort==4.2.5
kombu==4.0.1
Markdown==2.6.5
mongoengine==0.10.7
oauthlib==1.0.3
Pillow==2.8.1
#pkg-resources==0.0.0
psycopg2==2.6
pymongo==3.4.0
python-dateutil==2.5.3
python-magic==0.4.10
pytz==2016.10
redis==2.10.5
six==1.10.0
sorl-thumbnail==12.3
vine==1.1.3
